moved AMM class into amm.py
